requirements.txt
streamlit
tensorflow
pillow
numpy
requests
streamlit-lottie